+++ fractal/components/assessments/AssessmentScore/AssessmentScore.config.js
@@ -4,17 +4,8 @@ module.exports = {
     "collated": true,
     "preview": "@assessment-score-preview",
     "context": {
-        getScore: function() {
-            return new Promise((resolve, reject) => {
-                resolve({
-                    headerText: 'Lorem Ipsum',
-                    lastUpdated: '12/34/5678',
-                    percentFull: 75
-                });
-            });
-        },
-        getScoreSuccess: function(score) {
-            return score;
-        }
+        headerText: 'Lorem Ipsum',
+        lastUpdated: '12/34/5678',
+        score: 75
     }
 };