+++ fractal/components/breadcrumbs/Breadcrumb/Breadcrumb.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    "title": "Breadcrumb",
+    "status": "ready",
+    "collated": true,
+    "preview": "@breadcrumb-bar-preview",
+    "context": {
+        to: 'http://www.anywhere.com',
+        label: 'Lorem Ipsum A'
+    }
+};