+++ fractal/components/breadcrumbs/BreadcrumbBar/BreadcrumbBar.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+    "title": "Breadcrumb Bar",
+    "status": "ready",
+    "collated": true,
+    "preview": "@breadcrumb-bar-preview",
+    "context": {
+        breadcrumbs: [
+            {to: 'http://www.anywhere.com', label: 'Lorem Ipsum A'},
+            {to: 'http://www.somewhere.com', label: 'Lorem Ipsum B'},
+            {to: 'http://www.elsewhere.com', label: 'Lorem Ipsum C'}
+        ]
+    }
+};