+++ fractal/components/links/Link/Link.jsx
@@ -2,15 +2,16 @@ import React from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-const Link = ({ className, to, label, children, ...rest }) => {
+const Link = ({ className, labelClassName, to, label, children, ...rest }) => {
     const componentClass = 'c-link';
     const labelClass = `${componentClass}__label`;
     const componentClasses = classNames(componentClass, className);
+    const labelClasses = classNames(labelClass, labelClassName);
     const componentProps = { className: componentClasses, ...rest };
 
     return (
         <a href={to} {...componentProps} title={label}>
-            {label && <span className={labelClass}>{label}</span>}
+            {label && <span className={labelClasses}>{label}</span>}
             {children}
         </a>
     );
@@ -18,6 +19,7 @@ const Link = ({ className, to, label, children, ...rest }) => {
 
 Link.propTypes = {
     className: PropTypes.string,
+    labelClassName: PropTypes.string,
     children: PropTypes.node,
     to: PropTypes.string.isRequired,
     label: PropTypes.string